update existing query and test. delete duplicates
