+++ src/data/queries/index.ts
@@ -13,7 +13,7 @@ import * as AudienceTopics from './audienceTopics'
 import * as Alert from './alert'
 import * as AlertSingle from './alertSingle'
 import * as EmailContact from './emailContact'
-import * as BenefitsHub from './benefitsHub'
+import * as BenefitsHub from './benefitsHubLinks'
 import * as Wysiwyg from './wysiwyg'
 import * as StaticPathResources from './staticPathResources'
 import * as HeaderFooter from './headerFooter'
